chore: containers should always be killed by their parent, remove incorrect exits with 0 (#3595)
Previously if a container was killed through the agent/slot disable code path, it would kill containers directly rather than telling the allocation actor to kill itself. In all cases, a spurious container death will cause the allocation to die anyway and `det a --drain` is available when killing the allocation is not optimal (despite its issues, this will be true eventually). Moreover, a spurious container exit while the container is pulling causes the allocation actor to just receive an empty `containerStopped{}` which is indistinguishable from an exit 0. The allocation actor would interpret that exit incorrectly as a container deciding to exit early. 

This change makes kills always go through the allocation actor, so logic like counting towards max restarts is taken into account, and also fixes this bug since the allocation knows it was killed. Also a new container exit 'ContainerAborted' is added for cases where a container is killed before the agent can start it, rather than just returning 'container exited 0' since this is just wrong and could easily other issues.