chore: handle possible errors from docker daemon call /containers/cID/wait (#3592)
Though I can't see how from daemon code, the API make it appear possible for this Docker call to fail with an error the body. This changes our code to not ignore this error if it exists, since ignoring it would cause us to always return that the container exited with 0, causing upstream consumers to mistake it for a user requested stop (real exit 0).