fix: ensure allocation end time is always valid (#3705)
An edge case arose with our cluster heartbeat scheme where we save the
heartbeat, an allocation is started and the cluster dies.  In this case,
the end time will be set to the most recent cluster heartbeat but that
is before end time so the logs show `failed to aggregate resource
allocation: failed to add aggregate: ERROR: range lower bound must be
less than or equal to range upper bound (SQLSTATE 22000)
id="allocation-aggregator" system="master" type="allocationAggregator"`.