fix: trial rendezvous timeout should just warn (#3704)
In 0.17.0 (push arch) we had a regression such that failing to
rendezvous after 10m would cause a failure rather than just a warning.
This PR comprises 2 small changes.  First, the behavior is reverted back
to the old.  Second, the deadline does not start until the first
connection to avoid spurious warnings for things such as long start up
scripts.  Both are done rather than just the second since for k8s the
kill resulting from not suppressing the error could still be incorrect.