perf: rewrite get_experiments.sql for a better query plan (#3673)
Previously, the query for `get_experiments.sql` was something like: CTE
with filtered experiments, CTE to make page info from filtered
experiments, CTE to apply page info to filtered experiments (so,
filtered paginated experiments), then form the result from these CTEs.
The first CTE is very expensive because it does a lot of work that gets
thrown away by pagination; this change refactors the query to minimize
this wasted work.  This results in a bit of unfortunate code duplication
but I think this is fine.