test: add testing for cross-version master/agent clusters
For rolling upgrades to work properly, we need the master to be able to
communicate with agents from previous versions of Determined. (At least
for now, we're focusing on an upgrade strategy where the master is
upgraded first and the agents second.) And a first step toward ensuring
that is running tests on a cluster with mismatched versions, which this
adds using a custom devcluster config.