feat: remove global_batch_size requirement (#3724)
Previously, the master enforced that the global_batch_size
hyperparameter existed, but in the future we will have training loops
that do not require this hyperparameter (DeepspeedTrial, Core API).

In the experiment config schemas, this PR simply drops an old
constraint, allowing configurations that were not previously allowed.

In the master code, this removes any ability to attempt to read the
global_batch_size hyperparameter in go code.

In python code, removes .per_slot_batch_size and .global_batch_size from
the internal EnvContext object and moves
EnvContext._calculate_batch_sizes to det.util, where it can be called by
each training loop that still requires the global_batch_size
hyperparameter.  This implies that what used to be a configuration error
(caught before an experiment id was even assigned) is now a runtime
error.  This is unfortunate but unavoidable; it should never have been
the master's responsibility to enforce training loop inputs.