chore: add ProtoConverter (#3766)
"just write the struct as json and unmarshal it into protobuf" is not a
computationally free operation or an acceptable strategy overall.

Without some syntactic sugar though, coverting e.g. a
map[string]interface{} in go to a to a protobuf struct.Struct is
actually a pain because the conversion function can return an error.

ProtoConverter just captures and holds those errors until you've called
all of your conversion functions.