chore: add bun.DB singleton in master (#3769)
The bun.DB singleton will be the basis of future database code.

It was agreed that making the database connection a singleton is fine
because at runtime, because we only support one database anyway.  This
will allow for very concise and easy interfaces, like trials.List(...)
to list trials from the database.

When we want to unit test something like the trial actor (which makes a
few calls to the database), we can pass it a small interface to use, and
mock just that interface in the unit test.