fix: avoid double-timestamps in logs
We attach timestamps to all logs after-the-fact, so the python logging
library needs to be configured to not also emit timestamps.

Apparently, this was broken in distributed training since 0.17.0, and
only recently was it broken in non-distributed training as well.

The source of the bug was that there used to be two common paths by
which we configured logs, which I did not realize, and so in the push
architecture refactor I unwittingly took the worse one and used it on a
new code path.

This fix deletes that way to configure logs and uses the standard way
everywhere.