try again: use two separate enums this time
Having one enum was weird.  It did communicate the idea that the values
had similar mechanics, but the enum itself couldn't really be
documented, the functions which accepted it had to document how they
interpreted it.  Also, the searcher couldn't accept WorkersAskMaster so
it was further confusing.

The new system still communicates that the two enums have similar
mechanics by having them share enum values, but the documentation is
much easier to write and there typing also indicates accurately which
values are allowed.