HAL-2855 Enable multi-job runs with SLURM (#141)
* Enable multi-job runs with SLURM

Changes (some temporary) to enable experiments with multiple jobs to be launched by SLURM successfully.

* Enhance debug mode output for horovod initialization.
* Add per-SLURM-task /tmp directory creation so that work files do not conflict if concurrent jobs are writing to them.
  The directory is generated by mktemp and owned by the user and writable only by them.
* Configure an explicit (unknown) password hash in the shadow file instead of ! which disables the account.
* Add SLURM-specific log file initialization using /tmp to enable per-container/process logs to work with the shared file system.

* Drop generate tmp directory.

With SINGULARITY_CONTAIN=true, there is actually a local /tmp in the container that we can use.

Generalize the node/total args when a constraint is set to enable multi-job experiments on horizon which doesn't support --gpus allocation.
This stil needs more work so leaving the TODO.