feat: add core.DownloadMode (#3910)
In multi-worker settings, it is important that workers on the same node
share a single copy of a downloaded checkpoint.  On an 8-GPU compute
node, that amounts to an 8x bandwidth savings, and also it prevents
workers from stepping on each other's files.

This PR introduces DownloadMode.LocalWorkersShareDownload and
DownloadMode.NoSharedDownload settings, which apply to both
CheckpointContext.download() and CheckpointContext.restore_path().
This allows users to use .download() with or without this coordination,
in case they have special use cases that don't need coordination.

This PR also introduces unit tests for the CheckpointContext.