chore: demote _get_last_validation to internal
get_last_validation() had a bug where it might return a different value
before vs after you report any metrics.

Fixing it would require reconsidering how metrics are archived in the
master, which is a lot of work for an API that seems unlikely to be
useful to users, so we'll just make it internal.