fix: handle infinite validation metric values in more cases (#3992)
Since JSON can't represent infinity, we store infinite metric values as
the string `"Infinity"` or `"-Infinity"` in JSON in the database. Some
ways of taking those strings from PostgreSQL and converting them to
numbers work and preserve the infinite values, and some do not. The most
correct, and also most natural, way is to use `(map->>key)::float8`; the
similar `(map->key)::text::float8` adds quotes to string values and so
fails the last cast. (Also, we were using the `numeric` type before when
we did that, which does not support infinite values anyway.) Another
thing we were doing was running queries that returned jsonb values from
PostgreSQL back to Go and then attempting to do a standard
deserialization into a float64, which would fail on these string values;
we can take advantage of the same in-PostgreSQL cast as above to do the
right thing there.

We also need to add handling for when we have such an infinite value and
want to serialize it out to JSON. That's accomplished with a wrapper
type; for now, it's defined and used in the one place I've seen it to be
necessary so far, but it may make sense to move it somewhere and make it
public if there are other places it could be helpful.