fix: handle infinite metrics in searcher snapshots [DET-7122] (#3999)
This is simply applying the recently added "float extended with JSON
handling for infinite values" type to another place (and moving it to a
public place, and adding unmarshalling support, where it only had
marshalling before). See f915a695 for its introduction and other
context.

Possibly we should look into redoing marshalling entirely so we can add
support for infinities (and anything else that might require special
handling in the future) in just the one place rather than peppering this
type and type conversions everywhere. But that's going to be a whole
other thing.