chore: warn about ambiguous enum params (#3997)
the swagger spec doesn’t specify what enum type an operation needs and just specifies the members. we replace those with references to actual enums that are also defined in the spec but we run into cases where we can't tell which enum a member is referencing to. here we print a warning about those cases.