chore: Make MPIRUN work under horovodrun (#171)
We now have environment images with built-in OpenMPI.   When invoked the SLURM_JOBID variable
triggers integration with SLURM, however, we are running in a singularity container and SLRUM
may or may not have compatible configuration enabled.  We therefore clear the SLURM_JOBID variable
before invoking mpi so that mpirun will honor the args passed via horvod run to it describing
the hosts and process topology, otherwise mpi ends up wanting to launch all -np# processes on the local
causing an oversubscription error ("There are not enough slots available in the system").

Also fallback to OMPI_COMM_WORLD_RANK when HOROVOD_RANK is not defined.   This is not needed
for correct operation, but impacts [rank=#] logging prefix, so improves the user experience
and debugging.