chore: deprecate data layer
The Determined Data Layer has been fairly unsuccessful in attracting new
users, and by its nature it only works for a small number of users who
have small datasests that change infrequently.

The plan to remove it is to develop a satisfactory userspace alternative
that uses YogaDL directly rather than relying on the Data Layer feature
(Determined's integration with YogaDL).

Until then, remove docs, demote the examples to test fixtures, and
add deprecation warnings to discourage new users from using an API we
hope to delete in the future.