chore: HAL-2879 Cleanly shutdown all sshd servers on exit (#176)
Change the pidserver settings for sshd launches to SIGTERM during
pidserver exit.  Previously the setting was to wait for the sshd
processes, but they never exit so it can cause a hang.  With
this change when when the horovodrun command completes, and the
pidserver is about to exit, it instead sends SIGTERM to the
sshd processes and all containers can exit cleanly.