feat: add wrap_rank helper script
Core API users who use a hand-written or 3rd-party launcher will need to
prefix each log line with [rank=N] in order for the WebUI log viewer's
filter-by-rank feature to work.

Make this easier by exposing our internal wrapper script as a public
wrapper script.

The alternative of replacing sys.stdout with a class with a custom
write() method was deemed insufficient because many of our users depend
on C++ libraries that log information directly to the stdout file
descriptor.