chore: wait for process substition log filters [DET-6712] (#3930)
To prevent losing logs from very short commands in determined on 
kubernetes, all task container entrypoints are updated to wait on 
stdout and stderr process substitution filters before exiting. 
Additionally, the recommended fluentbit version is changed to 1.9.3, 
due to enhancements in SIGTERM handling. As a side effect from relying 
on so many shell features to fix this, we can no longer `exec` the 
container main process.