feat: allow NaN validation metrics [DET-7177] (#4150)
In several places, the harness had checks to filter out NaN validation
metric values, which previously could not be handled. This change
removes the checks, and testing has seemed to indicate that NaN values
are serialized and otherwise handled just fine without causing any
errors.

As a historical note, I found while tracing the history of this code
that NaN values were originally considered an error, then changed to not
being an error, then changed back again. The comment I changed was
actually incorrect until just now, as it applied to an older version of
the code.