refactor: pull in user service code changes from EE
For a while, the EE code has had some refactors in this code to expose
some functions for other packages, leading to occasional annoying
conflicts. Here we simply bring in those changes; this means there are
some functions that are unnecessarily exposed within the context of OSS
alone, which is mildly suboptimal but should be fine (and the changes
are otherwise normal-looking).