fix: allow `internal: null` for pre-0.15.6 experiments (#4197)
The internal field was removed with 0.18.0, along with the Native API.
But before 0.15.6, the `internal` field did not have an `omitempty` tag,
so it was still being stored in the database as `null`.  This causes
forking and certain REST API endpoints to fail on older experiments.

(cherry picked from commit 7f508e462fdfdf72db8eec71b4af1017f616b639)