chore: remove legacy code/docs for NCCL/Gloo port range config (#4187)
The port ranges used by NCCL and Gloo were previously configurable, but support
for this feature was removed quite a while ago. This commit removes the vestiges
of the previous feature, including docs and config parsing code. (We continue to
allow `nccl_port_range` and `gloo_port_range` to be specified in `master.yaml`
but those values are now ignored.)