fix: limit waiting for logger to 30 seconds
If the main process exits and the logger is still going after 30
seconds, assumes something is wrong and give up.

The 30 second window is configurabe by setting DET_LOG_WAIT_TIME in the
environment, in case of emergency.

An automated test is included to make sure that we log things after the
main process exits, but not after the DET_LOG_WAIT_TIME is exceeded.