chore: Merge Slurm-infrastructure support changes (#4376)
* HAL-2855 Enable multi-job runs with SLURM (#141)

* Enable multi-job runs with SLURM

Changes (some temporary) to enable experiments with multiple jobs to be launched by SLURM successfully.

* Enhance debug mode output for horovod initialization.
* Add per-SLURM-task /tmp directory creation so that work files do not conflict if concurrent jobs are writing to them.
  The directory is generated by mktemp and owned by the user and writable only by them.
* Configure an explicit (unknown) password hash in the shadow file instead of ! which disables the account.
* Add SLURM-specific log file initialization using /tmp to enable per-container/process logs to work with the shared file system.

* Drop generate tmp directory.

With SINGULARITY_CONTAIN=true, there is actually a local /tmp in the container that we can use.

Generalize the node/total args when a constraint is set to enable multi-job experiments on horizon which doesn't support --gpus allocation.
This stil needs more work so leaving the TODO.

* chore: FOUNDENG-20 deepspeed/cifar10_moe fails with no GPUs found! (#227)

* Drop master/static/srv/task-logging-setup.sh from merge

* chore: Make MPIRUN work under horovodrun (#171)

We now have environment images with built-in OpenMPI.   When invoked the SLURM_JOBID variable
triggers integration with SLURM, however, we are running in a singularity container and SLRUM
may or may not have compatible configuration enabled.  We therefore clear the SLURM_JOBID variable
before invoking mpi so that mpirun will honor the args passed via horvod run to it describing
the hosts and process topology, otherwise mpi ends up wanting to launch all -np# processes on the local
causing an oversubscription error ("There are not enough slots available in the system").

Also fallback to OMPI_COMM_WORLD_RANK when HOROVOD_RANK is not defined.   This is not needed
for correct operation, but impacts [rank=#] logging prefix, so improves the user experience
and debugging.

* HAL-2814 Provide data archive volumes to the dispatcher (#116)

* HAL-2814 Provide data archive volumes to the dispatcher

* Address feedback, and fix checkpoint_gc path support.

The checkpoint-gc archive has an cproto.RunArchive Path value that
that is not / like the others.   Regularize the root in the
generated launcher Archive argument so that we can treat them
consistently.

* chore: Merge Slurm-infrastructure support changes

One of many small changes coming to make the diffs more manageable.

* Revert "HAL-2814 Provide data archive volumes to the dispatcher (#116)"

This reverts commit bc0d4184f3fff71b17b3ac8418e27d988d374433.

* Drop scripts we need to refactor on dispatcher branch

* Incorporate review feedback on comments

* Add the enrich_task_logs.py as it is now referenced in an archive.

Add file and add to pod_test mocks.

* Switch from NCCLPortRange to GLOOPortRange

* Incorporate review feedback.

* Drop archive.go that has been merged to upstream already.

* Review feedback * works in place of BogusPasswordHash

* Run formatter.

* Improve grammar in comment.

* Another comment clarification.

* Run formatter

* Use improved variable filtering

* Migrated new threaded logging processing from ee

* Update comments

* Add try/finally

* Drop NCCL_PORT_RANGE

* fix: Drop tee to avoid random loss of enriched output lines

* Remove extra newline from Slurm logs

Co-authored-by: rcorujo <90728398+rcorujo@users.noreply.github.com>