feat: block incompatible agents from reconnecting [DET-7023]
In order to account for the possibility of the master resource pool
config changing between restarts and an old agent trying to reconnect,
we now have the master send the current resource pool config and the
agent send it back when it reconnects. If the master determines that the
previous configuration is incompatible with the current one, it
disconnects the agent.

The agent startup protocol now has `AgentStarted` containing only basic
information about the agent (not reattached containers) and coming
before `MasterSetAgentOptions`. If the master determines that the agent
is unacceptable, it disconnects it without updating any internal
state. Reattached containers were previously listed in the same message
but are now broken out to a later one, since the agent uses information
from the master to do the reattaching.

Some other logic that previously happened after `MasterSetAgentOptions`
now also happens before it; not strictly necessary, but it makes sense
to have the potentially disconnection-causing checks together and as
early as possible.