feat: docker auth improvements [DET-7633, DET-7636] (#4513)
Docker auth is improved in two ways. First if a serveraddress is specified in experiment config we will not send credentials to another server. The second is we now read the auths section of a users ~/.docker/config.json file meaning users can now use docker login on their agent nodes.