ci(test-e2e): move optional jobs into own workflow
Currently, all long-running jobs are only manually triggerable. We still
want this to be the case, but because of a lack of integration between
GH status checks and CircleCI pipeline reporting, we can't specify what
subset of the `test-e2e` job is okay not to run.

As a workaround, we're moving those manually-runnable jobs into a
separate workflow that'll be initiated the same as the current
`test-e2e` workflow, but will be marked as entirely optional for status
checks on PRs to the monorepo. This allows us to not have to circumvent
branch protections for PRs that don't require running those manual jobs.