ci(circleci): remove determined-ci context ref (#4613)
Using CircleCI's Contexts for use in our jobs makes them inaccessible
for execution in fork PRs. We really just want the values used in the
`determined-ci` context (DET_SEGMENT_*_KEY) to be accessible for use but
not scraping by web scrapers, so I've added those "secrets" as
environment variables for the `determined` project and removed the use
of the context.