fix: record operations at the right places around shutting down
The `s.Record` call was previously moved after the `if` block so that
the `Shutdown` operation was correctly recorded. However, this meant
that, running, e.g., random search with `max_concurrent_trials = 1`
would shut down after one trial, since the request to create the next
trial would not be counted in the check. So we simply have to record the
other operations first and then the `Shutdown` separately.