chore: disable positional args in bindings.py classes
generate_bindings.py generates deterministic code.  One way in which it
accomplishes that is by sorting parameters by name when it writes the
parameters to functions and methods.

That is safe to do as long as all args are keyword-only.  I made
keyword-only a requirement that for all the functions, but forgot to do
that for all the class __init__ methods.