chore: Use bindings.v1File instead of ContextItem (#4779)
This is an early part of a larger refactor.  Overall, the goal is to
move away from handcrafted API requests towards using the generated
bindings.

One blocker to that effort is that the creation of context directories
is definitely based on the handcrafted APIs.

Additionally, the context loading code has grown "crusty" over time; it
involved a lot of bandaid-style coversions between old and new formats,
in addition to the fact that a Context class existed, but it didn't
really do anything.  This PR focuses on migrating to a bindings-native
storage structure and reducing overall "crustiness".