test: Add default /etc/host bind mount in slurmcluster.sh (#395)
Podman V4.0 no longer maps /etc/hosts into the container which means that none of the admin/login nodes, nor compute node names can be resolved.  We are adding that to the default master.yaml configured on install with the launcher.  This adds the same /etc/host setup in slurmcluster.sh so that when using podman (-p) it maps /etc/hosts by default.