chore: promote Session to a first-class citizen (#4787)
Originally, Session was part of the python sdk, and not used anywhere
else.  Slowly it has come to be preferred for REST API interactions,
especially now that it is used with the new generated bindings.

Most things need bindings to work, and this requirement will grow
stronger as we port more and more code to use the generated bindings.
Bindings needs Session to work.  Session needs api.requests and
api.authentication to work.  This is all a big mess unless the importing
of Session is carefully controlled.

This PR promotes Session to be part of determined.common.api, with a
controlled import order to prevent failed circular imports.  Python
SDK-related import paths just point at the real definition now.

Additionally, this renames the old api.authentication.Session class to
UserTokenPair, to reduce confusion.