chore: generate python bindings with better types
The generated bindings are now generous about what they accept, but
strict in what they return (from a typing perspective).

In cases where an input type annotation (like Sequence) is converted to
an output type annotation (like List), the originally-provided value
will be used if it matches the output type, otherwise a new object of
the stricter type will be created.