chore: clean up det e create cli [MLG-44]
The original intention of experiment.py was a place for reusable code to
interact with the REST API.  The generated bindings now serve that
purpose, and experiment.py is just tech debt.

The contents of experiment.py met with the following fates:

- create_experiment: replaced with bindings
- create_experiment_and_follow_logs: deleted, it served no purpose
- create_test_experiment_and_follow_logs: deleted, it served no purpose
- patch_experiment: moved to cli/experiment.py, until all of its callers
  are replaced with bindings
- follow_experiment_logs: moved to cli/experiment.py where it is used
- follow_test_experiment_logs:  moved to cli/experiment.py where it is used
- generate_random_hparam_values: moved to _execution.py where it belongs
- make_test_experiment_config: moved to _execution.py where it belongs

Additionally, this change brings the /api/v1 experiment creation to
feature parity with the old echo endpoint, and migrates the CLI to use
the /api/v1 endpoint.

After this change, it should be possible to delete the old API endpoint,
though it might be prudent to wait a month or two for slow-to-update cli
users.