fix: Allow over-mounting of /tmp/work (FOUNDENG-205) (#410)
With Singularity /tmp is removed and re-linked to a user directory
to avoid the default host-wide share /tmp and provide more space that
the limited Singularity tmpfs space (10mb).  Make the removal of /tmp
handle injected sub-directories by bind mounts, by detecting the
error and reporting an ERROR message instead of failing.

Also add a FATAL error message giving context if an error in the shell
script is terminated due to non-zero exit (set -e).