fix: Initial comm failure should provide explaination (FDN-217)
Fix the exception raised when master is not reachable (Exception) -- APIHttpError only happens when call
completes without a successful status response could also be (SSLException, MasterNotFoundException).
This prevents the expected error message from being reported when the master is not reachable.

The initial prep_container was intended to fail reasonably quickly
to enable diagnostics of misconfiguration.   Recent changes
for re-using the common session (DET-8003) settings has caused the initial
communication to retry for more than 30 mins thus defeating the
original intent of prep_container.trial_prep and the error message
it provides.

This change lowers the session retry in prep_container (6 retries with 0.5
backoff -> 64 seconds) to enable the diagnostic message to be posted reasonably quickly.